requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
djangorestframework
requests==2.32.2
uuid
